+++ src/main/java/com/microtech/smartshop/dto/CommandeDTO.java
@@ -3,6 +3,8 @@ package com.microtech.smartshop.dto;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Builder;
 import lombok.Data;
+
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.List;
 
@@ -19,19 +21,19 @@ private Long clientId;
 private LocalDateTime date;
     @JsonProperty(access = JsonProperty.Access.READ_ONLY)
 
-private Double sousTotal;
+private BigDecimal sousTotal;
     @JsonProperty(access = JsonProperty.Access.READ_ONLY)
 
 private Double remise;
     @JsonProperty(access = JsonProperty.Access.READ_ONLY)
 
-private Double tva;
+private BigDecimal tva;
     @JsonProperty(access = JsonProperty.Access.READ_ONLY)
 
-private Double total;
+private BigDecimal total;
     @JsonProperty(access = JsonProperty.Access.READ_ONLY)
 
-private Double montantRestant;
+private BigDecimal montantRestant;
 private String codePromo;
     @JsonProperty(access = JsonProperty.Access.READ_ONLY)
